PERF: Re-use main app registry for raw HBS view lookups
da6edc1 introduced the `lookupView` method, which initialized a fresh resolver, and used it to directly look up raw-views (with no caching). This worked well, but is relatively slow.

This commit updates the `helperContext` to include access to the registry, and uses it to perform raw-view lookups. As well as re-using the registry, this also means we're making use of the resolver's built-in cache.

Co-authored-by: Ayke Halder <rr-it@users.noreply.github.com>